+++ data/src/main/java/com/mutualmobile/praxis/data/remote/JokeApiService.kt
@@ -1,6 +1,6 @@
 package com.mutualmobile.praxis.data.remote
 
-import com.mutualmobile.praxis.data.remote.model.NETJokeListData
+import com.mutualmobile.praxis.data.remote.model.JokeListEntity
 import retrofit2.Retrofit
 import retrofit2.http.GET
 
@@ -10,8 +10,10 @@ interface JokeApiService {
     fun createRetrofitService(retrofit: Retrofit): JokeApiService {
       return retrofit.create(JokeApiService::class.java)
     }
+
+    private const val GET_RANDOM_JOKES = "/jokes/random/5"
   }
 
-  @GET("/jokes/random/5")
-  suspend fun getFiveRandomJokes(): NETJokeListData
+  @GET(GET_RANDOM_JOKES)
+  suspend fun getFiveRandomJokes(): JokeListEntity
 }
