+++ domain/src/main/java/com/mutualmobile/praxis/domain/repository/IJokesRepo.kt
@@ -1,8 +1,8 @@
 package com.mutualmobile.praxis.domain.repository
 
 import com.mutualmobile.praxis.domain.SafeResult
-import com.mutualmobile.praxis.domain.model.DOMJokeList
+import com.mutualmobile.praxis.domain.model.JokeList
 
 interface IJokesRepo {
-    suspend fun getFiveRandomJokes(): SafeResult<DOMJokeList>
+    suspend fun getFiveRandomJokes(): SafeResult<JokeList>
 }
