+++ domain/src/main/java/com/mutualmobile/praxis/domain/usecases/BaseUseCase.kt
@@ -12,14 +12,31 @@ interface BaseUseCase<out Result, in ExecutableParam> {
    *
    * @return
    */
-  suspend fun perform(): Result = throw NotImplementedError()
+  fun perform(): Result = throw NotImplementedError()
 
   /**
-   * Perform an operation.
+   * Perform an operation with input parameters.
    *  Will throw an exception by default, if not implemented but invoked.
    *
    * @param params
    * @return
    */
-  suspend fun perform(params: ExecutableParam): Result = throw NotImplementedError()
+  fun perform(params: ExecutableParam): Result = throw NotImplementedError()
+
+  /**
+   * Perform an async operation with no input parameters.
+   * Will throw an exception by default, if not implemented but invoked.
+   *
+   * @return
+   */
+  suspend fun performAsync(): Result = throw NotImplementedError()
+
+  /**
+   * Perform an async operation.
+   *  Will throw an exception by default, if not implemented but invoked.
+   *
+   * @param params
+   * @return
+   */
+  suspend fun performAsync(params: ExecutableParam): Result = throw NotImplementedError()
 }
