+++ domain/src/main/java/com/mutualmobile/praxis/domain/usecases/GetFiveRandomJokesUseCase.kt
@@ -1,7 +1,7 @@
 package com.mutualmobile.praxis.domain.usecases
 
 import com.mutualmobile.praxis.domain.SafeResult
-import com.mutualmobile.praxis.domain.model.DOMJoke
+import com.mutualmobile.praxis.domain.model.Joke
 import com.mutualmobile.praxis.domain.repository.IJokesRepo
 
 /**
@@ -9,11 +9,11 @@ import com.mutualmobile.praxis.domain.repository.IJokesRepo
  */
 
 class GetFiveRandomJokesUseCase(private val jokesRepo: IJokesRepo) :
-    BaseUseCase<SafeResult<List<DOMJoke>>, Unit> {
+    BaseUseCase<SafeResult<List<Joke>>, Unit> {
 
-  override suspend fun perform(): SafeResult<List<DOMJoke>> {
+  override suspend fun performAsync(): SafeResult<List<Joke>> {
     return when (val result = jokesRepo.getFiveRandomJokes()) {
-      is SafeResult.Success -> SafeResult.Success(result.data.DOMJokes)
+      is SafeResult.Success -> SafeResult.Success(result.data.Jokes)
       is SafeResult.NetworkError -> result
       is SafeResult.Failure -> result
     }
