+++ featauthentication/src/main/java/com/praxis/authentication/common/model/FailureType.kt
@@ -1,0 +1,6 @@
+package com.praxis.authentication.common.model
+
+enum class FailureType(val message: String) {
+  EMAIL_NOT_VALID(message = "Email is not valid"),
+  PASSWORD_NOT_VALID(message = "Password should be at least 6 characters long")
+}
