+++ featauthentication/src/main/java/com/praxis/authentication/login/exceptions/FormValidationException.kt
@@ -1,0 +1,5 @@
+package com.praxis.authentication.login.exceptions
+
+import com.praxis.authentication.common.model.FailureType
+
+class FormValidationException(val failType: FailureType) : Throwable()
