+++ featauthentication/src/main/java/com/praxis/authentication/login/services/IValidate.kt
@@ -1,0 +1,5 @@
+package com.praxis.authentication.login.services
+
+interface IValidate {
+  fun validate(): Boolean
+}
