--- featauthentication/src/main/java/com/praxis/feat/authentication/services/Validatable.kt
@@ -1,5 +1,0 @@
-package com.praxis.feat.authentication.services
-
-interface Validatable {
-  fun validate(): Boolean
-}
