--- featauthentication/src/main/java/com/praxis/feat/authentication/ui/exceptions/FormValidationFailed.kt
@@ -1,5 +1,0 @@
-package com.praxis.feat.authentication.ui.exceptions
-
-import com.praxis.feat.authentication.ui.model.FailureType
-
-class FormValidationFailed(val failType: FailureType) : Throwable()
