--- featjokes/src/main/java/com/mutualmobile/feat/jokes/ui/home/InMemoryDataTemp.kt
@@ -1,8 +1,0 @@
-package com.mutualmobile.feat.jokes.ui.home
-
-import com.mutualmobile.feat.jokes.ui.model.UIJoke
-
-object InMemoryDataTemp {
-  // TODO dont do this in production lol :D
-  var jokes :List<UIJoke>? = null
-}
