--- featjokes/src/main/java/com/mutualmobile/feat/jokes/ui/home/JokeDetailsScreen.kt
@@ -1,58 +1,0 @@
-package com.mutualmobile.feat.jokes.ui.home
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.text.ClickableText
-import androidx.compose.material.*
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.withStyle
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import com.google.accompanist.insets.navigationBarsPadding
-import com.google.accompanist.insets.statusBarsPadding
-import com.mutualmobile.praxis.commonui.material.CommonTopAppBar
-import com.mutualmobile.praxis.commonui.theme.PraxisSurface
-import com.mutualmobile.praxis.commonui.theme.PraxisTheme
-
-@Composable
-fun JokeDetailsScreen(jokeDetailVM: JokeDetailVM = hiltViewModel()) {
-
-  Scaffold(
-    backgroundColor = PraxisTheme.colors.uiBackground,
-    contentColor = PraxisTheme.colors.textSecondary,
-    modifier = Modifier
-      .statusBarsPadding()
-      .navigationBarsPadding(),
-    topBar = {
-      CommonTopAppBar("Joke Detail")
-    }) {
-    PraxisSurface(
-      color = PraxisTheme.colors.uiBackground,
-      modifier = Modifier
-        .fillMaxHeight()
-        .fillMaxWidth(),
-    ) {
-      Column(
-        modifier = Modifier
-          .fillMaxHeight()
-          .fillMaxWidth()
-      ) {
-        val uiState by jokeDetailVM.uiState.collectAsState()
-        if (uiState is JokeDetailVM.UiState.SuccessState) {
-          Text(
-            text = (uiState as JokeDetailVM.UiState.SuccessState).joke.joke,
-            Modifier.padding(16.dp)
-          )
-        }
-      }
-    }
-  }
-}
