+++ featjokes/src/main/java/com/mutualmobile/jokes/home/InMemoryDataTemp.kt
@@ -1,0 +1,8 @@
+package com.mutualmobile.jokes.home
+
+import com.mutualmobile.jokes.model.JokeView
+
+object InMemoryDataTemp {
+  // TODO dont do this in production lol :D
+  var jokeViews :List<JokeView>? = null
+}
